NoticeBoard

+++ WebContent/view.jsp
@@ -7,7 +7,7 @@
 <html>
 <head>
 <meta http-equiv="Content-Type" content="text/html; charset= UTF-8">
-<meta name="viewport" content="width=divice-width" initial-scale="1"> <%--반응형 디자인--%>
+<meta name="viewport" content="width=divice-width", initial-scale="1"> <%--반응형 디자인--%>
 <link rel="stylesheet" href="css/bootstrap.css">
 <title>JSP게시판 웹 사이트</title>
 </head>
@@ -88,11 +88,11 @@
 				<tbody>
 					<tr>
 						<td style="width:20%">글 제목</td>
-						<td colspan="2"><%=bbs.getBbsTitle().replaceAll(" ","&nbsp;").replaceAll("<","&lt;").replaceAll("/n","<br>")%></td>
+						<td colspan="2"><%=bbs.getBbsTitle().replaceAll(" ","&nbsp;").replaceAll("<","&lt;").replaceAll(">","&gt;").replaceAll("/n","<br>")%></td>
 					</tr>
 					<tr>
 						<td>작성자</td>
-						<td colspan="2"><%=bbs.getUserID() %></td>
+						<td colspan="2"><%=bbs.getUserID().replaceAll(" ","&nbsp;").replaceAll("<","&lt;").replaceAll(">","&gt;").replaceAll("/n","<br>")%></td>
 					</tr>
 					<tr>
 						<td>작성 일자</td>
@@ -102,7 +102,7 @@
 					</tr>
 					<tr>
 						<td>내용</td>
-						<td colspan="2" style="min-height:200px;text-align:left;"><%=bbs.getBbsContent().replaceAll(" ","&nbsp;").replaceAll("<","&lt;").replaceAll("/n","<br>") %></td>
+						<td colspan="2" style="min-height:200px;text-align:left;"><%=bbs.getBbsContent().replaceAll(" ","&nbsp;").replaceAll("<","&lt;").replaceAll(">","&gt;").replaceAll("/n","<br>") %></td>
 					</tr>
 				</tbody>
 			</table>
@@ -111,7 +111,7 @@
 				if(userID !=null &&userID.equals(bbs.getUserID())){//현재 사용자가 작성자와 같다면 ↓
 			%>
 				<a href="update.jsp?bbsID=<%=bbsID %>"class="btn btn-primary">수정</a>		   <!--이부분 출력 -->
-				<a href="deleteAction.jsp?bbsID=<%=bbsID %>"class="btn btn-primary">삭제</a>	
+				<a onclick="return confirm('정말 삭제하시겠습니까?')" href="deleteAction.jsp?bbsID=<%=bbsID %>"class="btn btn-primary">삭제</a>	
 			<% 		
 				}
 			%>
